% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boliga_all_houses.R
\name{boliga_all_houses}
\alias{boliga_all_houses}
\title{Get all houses within a postal code.}
\usage{
boliga_all_houses(postal_code)
}
\arguments{
\item{postal_code}{Postal code.}
}
\description{
The function takes the search results from the base url
and iterates until there's no "next" page.
}
