% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boliga_bbr_get_single.R
\name{boliga_bbr_get_single}
\alias{boliga_bbr_get_single}
\title{Function to get the BBR content for a specific relative url.}
\usage{
boliga_bbr_get_single(relative_url)
}
\arguments{
\item{relative_url}{The relative url.}
}
\description{
Function to get the BBR content for a specific relative url.
}
\examples{

library(boliga)
re_url <- "/bbrinfo/3DC48AAF-5E35-43A7-BCB3-3A34565291F6"
boliga_bbr_get_single(re_url)

}
