% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boliga_clean_address.R
\name{boliga_clean_address}
\alias{boliga_clean_address}
\title{This is a helper function to take the base adress, which concatenate
the house number and postal code, and return a new and improved adress
seperating the house number and postal code.}
\usage{
boliga_clean_address(bad_address)
}
\arguments{
\item{bad_adress}{Character vector.}
}
\description{
This is a helper function to take the base adress, which concatenate
the house number and postal code, and return a new and improved adress
seperating the house number and postal code.
}

