% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boliga_create_base_url.R
\name{boliga_create_base_url}
\alias{boliga_create_base_url}
\title{Function to create the base url for webscraping boliga.}
\usage{
boliga_create_base_url(min_sale_date = NULL, max_sale_date = Sys.Date(),
  type = c("Alle", "Villa", "Rækkehus", "Ejerlejlighed", "Fritidshus",
  "Landejendom"), postal_code = NULL)
}
\arguments{
\item{min_sale_date}{as.Date(). The minimum sale date.}

\item{max_sale_date}{as.Date(). The maximum sale date.}

\item{type}{type "Alle", "Villa", "Rækkehus", "Ejerleglighed", "Fritidshus", "Landejendom"}

\item{post_no}{Post number.}
}
\description{
Function to create the base url for webscraping boliga.
}
\examples{

library(boliga)

boliga_create_base_url(min_sale_date = Sys.Date() - 100,
                       max_sale_date = Sys.Date(),
                       type = "Fritidshus",
                       postal_code = 4500)
                       
}
