% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boliga_get_table.R
\name{boliga_get_table}
\alias{boliga_get_table}
\title{This function retuns a list with the search results
from a specific url}
\usage{
boliga_get_table(url_address)
}
\arguments{
\item{url_address}{The url you wan't to start}
}
\description{
This function retuns a list with the search results
from a specific url
}
