% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boliga_webscrape_sold.R
\name{boliga_webscrape_sold}
\alias{boliga_webscrape_sold}
\title{This is the primary function to webscrape boliga
for the actual sale prices}
\usage{
boliga_webscrape_sold(min_sale_date, max_sale_date, type, postal_code)
}
\arguments{
\item{min_sale_date}{Sale date from.}

\item{max_sale_date}{Sale date to.}

\item{type}{Type.}

\item{postal_code}{Postal code.}
}
\description{
The function takes the search results from the base url
and iterates until there's no "next" page.
}
