% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boliga.R
\name{load_files}
\alias{load_files}
\title{Load CSV files in working directory}
\usage{
load_files(concat = T)
}
\arguments{
\item{concat}{Should the tibbles be concatenated? Default is FALSE and it returns a list.}
}
\description{
To plot the mean time on market by dates when data is scraped do something like this:
#ggplot(dfs_clean, aes(x = today, y = time_on_market)) + stat_summary(geom = "line", fun.y = "mean", fun.args = list(na.rm = T))
}
