% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boliga.R
\name{scrape_all_pages}
\alias{scrape_all_pages}
\title{Scrape pages silently}
\usage{
scrape_all_pages()
}
\value{
A list with four elements. The results slot contains the output from iterations without failure. Check errors for scraped pages where code failed.
}
\description{
Scrape pages silently
}
