% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boliga.R
\name{scrape_all_pages_silently}
\alias{scrape_all_pages_silently}
\title{Scrape pages silently}
\usage{
scrape_all_pages_silently()
}
\value{
A list with four elements. The results slot contains the output from iterations without failure. Check warnings slot for possible problems for scraped pages.
}
\description{
Scrape pages silently
}
