% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boliga.R
\name{scrape_page}
\alias{scrape_page}
\title{Scrape a single page from boliga}
\usage{
scrape_page(page_num)
}
\arguments{
\item{page_num}{'page_num' should be an integer and is pasted to &page=}
}
\value{
a tibble with nine columns.
}
\description{
The function will scrape a given page by providing the page number as argument.
}
