% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boliga_webscrape_sold.R
\name{boliga_webscrape_sold}
\alias{boliga_webscrape_sold}
\title{This is the primary function to webscrape boliga
for the actual sale prices}
\usage{
boliga_webscrape_sold(url_address, random_pause = TRUE)
}
\arguments{
\item{url_address}{The url you wan't to start}

\item{pause}{the function with rnorm(1, 1, 0.2) if TRUE, when webscrapping.}
}
\description{
The function takes the search results from the base url
and iterates until there's no "next" page.
}

